Parser Fix
Moving in the direction of just simply either downloading or logging in.
Giving the user the choice of either downloading a track or playlist seems
redundant, we can just simply check the url itself to determine whether or not
it's a track or playlist. Less options the better.

Also changed the parser as a result, we simply pass in the URLs we wanna
download or just call 'cortez.py login', and we can then do the entire flow
or check to see if the user already has an access token.