Major Refactoring
Seperated the main cortez downloading function into it's own seprate class,
along with our config.py file. Instead of just reading directly for the config
values, we create a class and check to see if there are any breaking changes,
then we pass the class along for a simpler method of retrieving any variables.