Config Alterations
Started a rough outline of the ability to edit the config file
via the terminal. Still a decent amount of work to be done,
in particular an error prompt when invalid information is attempted
to be saved, like whether or not to save an auth token.

Also fixed an issue with the downloader when attempting to download
songs which don't have a direct download link nor a stream URL.
Apparently songs can be set private which prohibit, or just make more
difficult, to get and download the stream version of the song.