Config Additions
Still much work to be done, but updated the config to display
a new window whenever a user attempts to save their config. It
isn't smart yet in the sense that it doesn't know when something
has been edited nor is it pretty; much work left to be done
to properly format buttons and such.